+++ stock_api/Service/ValueObject/WarehouseProductVo.cs
@@ -5,6 +5,6 @@ namespace stock_api.Service.ValueObject
     public class WarehouseProductVo : WarehouseProduct
     {
         public string? AnotherUnit {  get; set; }
-        public int? AnotherUnitConversion { get; set; }
+        public float? AnotherUnitConversion { get; set; }
     }
 }
